Concurrency, Loading, Better Responsiveness, Catching Exceptions from Query
Program no longer blocks while results are being queried (300ms per query).
Query starts as a new thread, and a JFXSpinner takes the place of the table.
style.css updated for the spinner
Also catching exceptions in case no result is found, but these shouldn't be thrown in the first place